Aggiunta eliminazione playlist + modifica login
-Aggiunta la chiamata per l'eliminazione della playlist
-Modifica alla visualizzazione delle activity se l'utente si è già autenticato in precedenza

+++ SoundLab/app/src/main/java/com/soundlab/app/view/fragment/AddToPlaylistFragment.java
@@ -54,7 +54,7 @@ public class AddToPlaylistFragment extends Fragment {
                 // TODO: Carica le playlist dal backend
 
                 // Aggiungi le playlist
-                playlistArrayList.add(new Playlist(1, "Playlist1", "Rock",false, null));
+                playlistArrayList.add(new Playlist(1L, "Playlist1", "Rock",false, null));
 
                 // Inizializza l'adapter e passa la lista di playlist
                 AddToPlaylistAdapter addToPlaylistAdapter = new AddToPlaylistAdapter(this, playlistArrayList, song);
@@ -108,7 +108,7 @@ public class AddToPlaylistFragment extends Fragment {
 
                 // TODO: Inserire la playlist nel backend ed ottenere l'id.
 
-                int idPlaylist = 1; // Da cambiare con l'id ottenuto nel backend
+                Long idPlaylist = 1L; // Da cambiare con l'id ottenuto nel backend
                 Playlist newPlaylist = new Playlist(idPlaylist, nome_playlist, genere_playlist, false, null);
 
                 AddToPlaylistAdapter addToPlaylistAdapter = (AddToPlaylistAdapter) recyclerView.getAdapter();
